refactor(cost-explorer): replace CostAdvancedSettingsContainer with AdminAdvancedSettingsPage
- Removed CostAdvancedSettingsContainer.vue component.
- Added AdminAdvancedSettingsPage.vue component for advanced settings management.
- Updated routes to reflect the new component structure.
