+++ apps/web/src/services/cost-explorer/routes/admin/routes.ts
@@ -16,15 +16,11 @@ import { DYNAMIC_COST_QUERY_SET_PARAMS, MANAGED_COST_QUERY_SET_IDS } from '@/ser
 import CostExplorerHome from '@/services/cost-explorer/pages/CostExplorerHome.vue';
 import { ADMIN_COST_EXPLORER_ROUTE } from '@/services/cost-explorer/routes/admin/route-constant';
 
-const CostAdvancedSettingsContainer = () => import('@/services/cost-explorer/components/CostAdvancedSettingsContainer.vue');
-
 const CostExplorerContainer = () => import('@/services/cost-explorer/CostExplorerContainer.vue');
 
+const AdminAdvancedSettingsPage = () => import('@/services/cost-explorer/pages/admin/AdminAdvancedSettingsPage.vue');
 const AdminBudgetMainPage = () => import('@/services/cost-explorer/pages/admin/AdminBudgetMainPage.vue');
-// const AdminBudgetCreatePage = () => import('@/services/cost-explorer/pages/admin/AdminBudgetCreatePage.vue');
 const AdminBudgetDetailPage = () => import('@/services/cost-explorer/pages/admin/AdminBudgetDetailPage.vue');
-const AdminDomainSettingsAnomalyDetectionConfigurationPage = () => import('@/services/cost-explorer/pages/admin/AdminAdvancedSettingsAnomalyDetectionConfigurationPage.vue');
-const AdminDomainSettingsCurrencyConverterPage = () => import('@/services/cost-explorer/pages/admin/AdminAdvancedSettingsCurrencyConverterPage.vue');
 
 const AdminCostAnalysisPage = () => import('@/services/cost-explorer/pages/admin/AdminCostAnalysisPage.vue');
 const CostReportPage = () => import('@/services/cost-explorer/pages/CostReportPage.vue');
@@ -213,24 +209,7 @@ const adminCostExplorerRoutes: RouteConfig = {
                 menuId: MENU_ID.COST_ADVANCED_SETTINGS,
                 translationId: MENU_INFO_MAP[MENU_ID.COST_ADVANCED_SETTINGS].translationId,
             },
-            redirect: () => ({
-                name: ADMIN_COST_EXPLORER_ROUTE.COST_ADVANCED_SETTINGS.ANOMALY_DETECTION_DOMAIN_CONFIGURATION._NAME,
-            }),
-            component: CostAdvancedSettingsContainer,
-            children: [
-                {
-                    path: 'anomaly-detection-configuration',
-                    name: ADMIN_COST_EXPLORER_ROUTE.COST_ADVANCED_SETTINGS.ANOMALY_DETECTION_DOMAIN_CONFIGURATION._NAME,
-                    meta: { lsbVisible: true, translationId: MENU_INFO_MAP[MENU_ID.ANOMALY_DETECTION_DOMAIN_CONFIGURATION].translationId, menuId: MENU_ID.ANOMALY_DETECTION_DOMAIN_CONFIGURATION },
-                    component: AdminDomainSettingsAnomalyDetectionConfigurationPage,
-                },
-                {
-                    path: 'currency-converter',
-                    name: ADMIN_COST_EXPLORER_ROUTE.COST_ADVANCED_SETTINGS.CURRENCY_CONVERTER._NAME,
-                    meta: { lsbVisible: true, translationId: MENU_INFO_MAP[MENU_ID.CURRENCY_CONVERTER].translationId, menuId: MENU_ID.CURRENCY_CONVERTER },
-                    component: AdminDomainSettingsCurrencyConverterPage,
-                },
-            ],
+            component: AdminAdvancedSettingsPage as any,
         },
     ],
 };
